Flytt dto til egen fil
